updated focus claim fixed a bug with sps unstake fixed a bug with backup option for finding ongoing battle updated chest tier information for the scenario where api returns null value removed CP from advance league

+++ SplinterlandsRObot/Player/PlayerSeason/SeasonProgress.cs
@@ -9,8 +9,8 @@
 
             seasonChests = CalculateEarnedChests(
                     Math.Max(
-                        modern != null ? modern.chest_tier : 0,
-                        wild != null ? wild.chest_tier : 0
+                        modern != null ? (int)modern.chest_tier : 0,
+                        wild != null ? (int)wild.chest_tier : 0
                         ),
                     totalRshars
                     );
